Custom configuration file can be passed to ReFrame (yet, some unittests still fail)
