+++ unittests/fixtures.py
@@ -3,10 +3,9 @@
 #
 import os
 
+import reframe.frontend.config as config
 from reframe.core.modules import (get_modules_system,
                                   init_modules_system, NoModImpl)
-from reframe.frontend.loader import autodetect_system, SiteConfiguration
-from reframe.settings import settings
 
 TEST_RESOURCES = os.path.join(
     os.path.dirname(os.path.realpath(__file__)), 'resources')
@@ -83,9 +82,10 @@ def init_native_modules_system():
 
 
 # Guess current system and initialize its modules system
-_site_config = SiteConfiguration()
+settings = config.load_from_file("reframe/settings.py")
+_site_config = config.SiteConfiguration()
 _site_config.load_from_dict(settings.site_configuration)
-HOST = autodetect_system(_site_config)
+HOST = config.autodetect_system(_site_config)
 init_native_modules_system()
 
 
@@ -95,7 +95,7 @@ def get_test_config():
     Returns a tuple of system, partition and environment that you can pass to
     `RegressionTest`'s setup method.
     """
-    site_config = SiteConfiguration()
+    site_config = config.SiteConfiguration()
     site_config.load_from_dict(TEST_SITE_CONFIG)
 
     system = site_config.systems['testsys']
